Revert "Shade cglib"
This reverts commit dcf55993df228f29a3a6f43502d0bd0139ebaabd.

It turns out, cglib breaks when you shade it (because it hardcodes class names
inside strings). Luckily, for whatever reason, the shading didn't work in
jitpack's builds, so this only caused an issue in local builds - mods that
fetched the library from jitpack (which should be all of them) were unaffected.