Make SSLHacker work even if we're not the first to connect with HTTPS
This caused an issue between CraftPresence-2.0.0-rc.3+1.7.10.jar and
Et_Futurum_Requiem-2.4.6.jar on Java 8u51. CP connects to the internet on a
thread, therefore it was possible for it to initialize the SSL socket factory
before Et Futurum hacks it.

Now the hacking will work regardless of when it runs.